fixes: QoL fixes
NOTES:
- No major fixes, just making the program slightly more tuned, adhering to better practices

PLANS:
- I want to have a more defined sepreration between math logic and UI / GUI logic, simulation needs to take frm the mathematical and physical classes in such a way that they never overlap!

+++ Sim_Engine/EngineLib/src/Rendering/IBL.cpp
@@ -319,8 +319,8 @@ namespace render {
 
 		const unsigned int MAX_MIP_LEVELS = 5;
 		for (unsigned int mip = 0; mip < MAX_MIP_LEVELS; ++mip) {
-			unsigned int mipWidth = PREFILTER_RES * std::pow(0.5, mip);
-			unsigned int mipHeight = PREFILTER_RES * std::pow(0.5, mip);
+			unsigned int mipWidth = PREFILTER_RES * (int)std::pow(0.5, mip);
+			unsigned int mipHeight = PREFILTER_RES * (int)std::pow(0.5, mip);
 			
 			glBindRenderbuffer(GL_RENDERBUFFER, captureRBO);
 			glRenderbufferStorage(GL_RENDERBUFFER, GL_DEPTH_COMPONENT24, mipWidth, mipHeight);
